moved the shared file inside msm as the da-title is not using it anymore
